+++ packages/coding-bootcamp-testimonials-slider/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './controls';
+export * from './image';
+export * from './quote';